django_backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
sqlparse==0.5.3
typing_extensions==4.15.0
psycopg2-binary==2.9.9
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
channels==4.0.0
daphne==4.0.0
django-ninja==1.4.5 
django-ninja-extra==0.30.2
django-ninja-jwt==5.4.0
django-ninja-simple-jwt==0.6.1
django-cors-headers==4.3.1
